tools/downloader: split off the config loading functionality into a module
