+++ tools/downloader/src/open_model_zoo/model_tools/quantizer.py
@@ -22,7 +22,9 @@ from pathlib import Path
 
 import yaml
 
-from open_model_zoo.model_tools import _common, _reporting
+from open_model_zoo.model_tools import (
+    _configuration, _common, _reporting,
+)
 
 DEFAULT_POT_CONFIG_BASE = {
     'compression': {
@@ -141,7 +143,7 @@ def main():
             sys.exit('Unable to locate Post-Training Optimization Toolkit. '
                 + 'Use --pot or run setupvars.sh/setupvars.bat from the OpenVINO toolkit.')
 
-    models = _common.load_models_from_args(parser, args)
+    models = _configuration.load_models_from_args(parser, args)
 
     # We can't mark it as required, because it's not required when --print_all is specified.
     # So we have to check it manually.
